requirements/base.txt
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
django-filter==23.5
python-decouple==3.8
Pillow==10.4.0
gunicorn==21.2.0
